# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: StudyCenter API
  version: 0.1.0
  description: Contract for the StudyCenter learning-space and knowledge-card API.
servers:
  - url: /api/v1
paths:
...
  /learning-spaces:
    post:
      operationId: createLearningSpace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLearningSpaceRequest'
      responses:
        '202':
          description: Learning space queued for asynchronous generation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningSpace'
    get:
      operationId: listLearningSpaces
      responses:
        '200':
          description: Learning spaces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningSpaceList'
  /course-design/turn:
    post:
      operationId: courseDesignTurn
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CourseDesignTurnRequest' }
      responses:
        '200':
          description: Next course intake question or ready preview
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseDesignTurnResponse' }
  /course-design/outline:
    post:
      operationId: courseDesignOutline
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CourseOutlineRequest' }
      responses:
        '200':
          description: Generated course outline
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseOutlineResponse' }
        '502':
          description: AI did not return a valid outline
  /course-design/outline/revise:
    post:
      operationId: reviseCourseDesignOutline
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CourseOutlineRevisionRequest' }
      responses:
        '200':
          description: Revised course outline
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseOutlineRevisionResponse' }
        '502':
          description: AI did not return a valid outline
  /learning-spaces/{spaceId}:
    parameters:
      - $ref: '#/components/parameters/SpaceId'
    get:
      operationId: getLearningSpace
      responses:
        '200':
          description: Learning space
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningSpace'
        '404':
          $ref: '#/components/responses/NotFound'
  /learning-spaces/{spaceId}/generation:
    parameters:
      - $ref: '#/components/parameters/SpaceId'
    get:
      operationId: getCourseGenerationStatus
      responses:
        '200':
          description: Course generation status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationStatus'
    put:
      operationId: retryCourseGeneration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLearningSpaceRequest'
      responses:
        '200':
          description: Failed course generation queued again
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningSpace'
        '409':
          description: Course generation is not eligible for retry
  /learning-spaces/{spaceId}/runtime:
    parameters:
      - $ref: '#/components/parameters/SpaceId'
    get:
      operationId: getLearningRuntime
      responses:
        '200':
          description: Last persisted learning position and branch return path
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LearningRuntime'
                  - type: 'null'
    put:
      operationId: updateLearningRuntime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLearningRuntimeRequest'
      responses:
        '200':
          description: Persisted learning runtime
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningRuntime'
...
components:
  parameters:
    SpaceId:
      name: spaceId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    CreateLearningSpaceRequest:
      type: object
      required: [title, learningGoal]
      properties:
        title: { type: string, minLength: 1 }
        learningGoal: { type: string, minLength: 1 }
        courseBrief: { $ref: '#/components/schemas/CourseBrief' }
        courseScale: { type: string, enum: [quick, standard, series], default: standard }
        courseOutline: { type: array, maxItems: 12, items: { $ref: '#/components/schemas/CourseOutlineItem' } }
    CourseBrief:
      type: object
      properties:
        topic: { type: string }
        learningOutcome: { type: string }
        priorKnowledge: { type: string }
        useCase: { type: string }
        focus: { type: array, items: { type: string } }
        excludedTopics: { type: array, items: { type: string } }
        preferredStyle: { type: array, items: { type: string } }
        timeBudgetMinutes: { type: integer, nullable: true }
    CourseIntakeMessage:
      type: object
      required: [role, content]
      properties:
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
    CourseOutlineItem:
      type: object
      required: [title, objective]
      properties:
        title: { type: string }
        objective: { type: string }
    CourseDesignTurnRequest:
      type: object
      properties:
        messages: { type: array, items: { $ref: '#/components/schemas/CourseIntakeMessage' } }
        brief: { $ref: '#/components/schemas/CourseBrief' }
        courseScale: { type: string, enum: [quick, standard, series] }
        skip: { type: boolean, default: false }
    CourseDesignTurnResponse:
      type: object
      required: [brief, assistantMessage, ready, recommendedScale, courseScale, outline, turn]
      properties:
        brief: { $ref: '#/components/schemas/CourseBrief' }
        assistantMessage: { type: string }
        question: { type: string, nullable: true }
        quickOptions: { type: array, items: { type: string } }
        ready: { type: boolean }
        recommendedScale: { type: string, enum: [quick, standard, series] }
        courseScale: { type: string, enum: [quick, standard, series] }
        outline: { type: array, items: { $ref: '#/components/schemas/CourseOutlineItem' } }
        turn: { type: integer, minimum: 0, maximum: 3 }
    CourseOutlineRequest:
      type: object
      required: [brief, courseScale]
      properties:
        brief: { $ref: '#/components/schemas/CourseBrief' }
        courseScale: { type: string, enum: [quick, standard, series] }
    CourseOutlineResponse:
      type: object
      required: [courseScale, outline]
      properties:
        courseScale: { type: string, enum: [quick, standard, series] }
        outline: { type: array, minItems: 1, maxItems: 12, items: { $ref: '#/components/schemas/CourseOutlineItem' } }
    CourseOutlineRevisionMessage:
      type: object
      required: [role, content]
      properties:
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
    CourseOutlineRevisionRequest:
      type: object
      required: [brief, courseScale, currentOutline, feedback]
      properties:
        brief: { $ref: '#/components/schemas/CourseBrief' }
        courseScale: { type: string, enum: [quick, standard, series] }
        currentOutline: { type: array, minItems: 1, maxItems: 12, items: { $ref: '#/components/schemas/CourseOutlineItem' } }
        feedback: { type: string }
        messages: { type: array, maxItems: 20, items: { $ref: '#/components/schemas/CourseOutlineRevisionMessage' } }
    CourseOutlineRevisionResponse:
      type: object
      required: [courseScale, outline, assistantMessage]
      properties:
        courseScale: { type: string, enum: [quick, standard, series] }
        outline: { type: array, minItems: 1, maxItems: 12, items: { $ref: '#/components/schemas/CourseOutlineItem' } }
        assistantMessage: { type: string }
    LearningSpace:
      type: object
      required: [id, title, learningGoal, rootCardId, generationStatus, generationPhase, courseBrief, courseScale, createdAt]
      properties:
        id: { type: string }
        title: { type: string }
        learningGoal: { type: string }
        rootCardId: { type: string, nullable: true }
        generationStatus: { type: string, enum: [queued, running, completed, failed] }
        generationPhase: { type: string, enum: [queued, generating, saving, completed, failed] }
        generationError: { type: string, nullable: true }
        courseBrief: { $ref: '#/components/schemas/CourseBrief' }
        courseScale: { type: string, enum: [quick, standard, series] }
        courseOutline: { type: array, items: { $ref: '#/components/schemas/CourseOutlineItem' } }
        createdAt: { type: string, format: date-time }
    GenerationStatus:
      type: object
      required: [spaceId, status, phase, rootCardId, updatedAt]
      properties:
        spaceId: { type: string }
        status: { type: string, enum: [queued, running, completed, failed] }
        phase: { type: string, enum: [queued, generating, saving, completed, failed] }
        error: { type: string, nullable: true }
        rootCardId: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time }
    LearningSpaceList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/LearningSpace' }
        nextCursor: { type: string, nullable: true }
...
    LearningNavigationEntry:
      type: object
      required: [cardId]
      properties:
        cardId: { type: string }
        sectionId: { type: string, nullable: true }
    UpdateLearningRuntimeRequest:
      type: object
      required: [currentCardId, navigationStack]
      properties:
        currentCardId: { type: string }
        currentSectionId: { type: string, nullable: true }
        navigationStack:
          type: array
          items: { $ref: '#/components/schemas/LearningNavigationEntry' }
        eventType: { type: string, default: navigation }
    LearningRuntime:
      type: object
      required: [id, spaceId, currentCardId, navigationStack, updatedAt]
      properties:
        id: { type: string }
        spaceId: { type: string }
        currentCardId: { type: string, nullable: true }
        currentSectionId: { type: string, nullable: true }
        sourceCardId: { type: string, nullable: true }
        sourceSectionId: { type: string, nullable: true }
        navigationStack:
          type: array
          items: { $ref: '#/components/schemas/LearningNavigationEntry' }
        updatedAt: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }
            requestId: { type: string }